Remove GoalReachedCondition test
