+++ nav2_behavior_tree/test/plugins/condition/CMakeLists.txt
@@ -5,7 +5,3 @@ ament_target_dependencies(test_condition_distance_traveled ${dependencies})
 ament_add_gtest(test_condition_time_expired test_time_expired.cpp)
 target_link_libraries(test_condition_time_expired nav2_time_expired_condition_bt_node)
 ament_target_dependencies(test_condition_time_expired ${dependencies})
-
-ament_add_gtest(test_condition_goal_reached test_goal_reached.cpp)
-target_link_libraries(test_condition_goal_reached nav2_goal_reached_condition_bt_node)
-ament_target_dependencies(test_condition_goal_reached ${dependencies})
